+++ test/fixtures/cookbooks/divvy_test/recipes/mac_direct_download.rb
@@ -1,5 +1,5 @@
 # Encoding: UTF-8
 
 divvy_app 'default' do
-  provider Chef::Provider::DivvyApp::MacOsX::Direct
+  source :direct
 end